добавлена разметка треков

+++ DataBase/db_models/Audio.cs
@@ -108,10 +108,10 @@ public class DbMood(int id, string name, string? description)
 public class DbGenre(int id, string name, string? description)
     : DbAudioParameterValue(id, DbAudioParameter.GenreParameter.Id, name, description);
 
-public class DbUser(int id, int chatId, string username)
+public class DbUser(int id, long chatId, string username)
 {
     public int Id = id;
-    public int ChatId = chatId;
+    public long ChatId = chatId;
     public string Username = username;
 }
